attachment 改用 unsigned int...这个上次怎么忘了 ft 代码里有的地方写 sizeof(int)，有些地方写 sizeof(unsigned int)...
git-svn-id: http://svn.kcn.cn/repos/kbs/trunk@7944 de4889a5-8ad4-4bf3-bd87-57997c21d515

+++ kbs_bbs/site/default.h
@@ -666,7 +666,7 @@ typedef struct fileheader {     /* This structure is used to hold data in */
     char owner[OWNER_LEN];
     unsigned int eff_size; /* 版面文章: 表示有效字节数，信件: 表示文件大小 - atppp */
     int posttime;
-    long attachment;
+    unsigned int attachment;
     char title[ARTICLE_TITLE_LEN];
     unsigned char accessed[4];
 } fileheader;
